add root dir
