+++ src/app/root/app.module.ts
@@ -3,10 +3,10 @@ import { BrowserModule } from '@angular/platform-browser';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
-import { MaterialModule } from './modules/material/material.module';
+import { MaterialModule } from '../modules/material/material.module';
 import { HeaderComponent } from './header/header.component';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
-import { CommonService } from './services/common.service';
+import { CommonService } from '../services/common.service';
 import { FooterComponent } from './footer/footer.component';
 
 @NgModule({
