+++ src/app/root/footer/footer.component.scss
