+++ src/app/root/footer/footer.component.ts
